Restart refactor by first deprecating Sheet.
