+++ packages/plugins/plugin-sheet/src/components/SheetContainer/index.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+import { SheetContainer } from './SheetContainer';
+
+export default SheetContainer;
