+++ packages/plugins/plugin-sheet/src/components/SheetContext/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './SheetContext';
