+++ packages/plugins/plugin-sheet/src/components/index.ts
@@ -5,7 +5,7 @@
 import React from 'react';
 
 export * from './ComputeGraph';
-export * from './Sheet';
+export * from './GridSheet';
 
 // Lazily load components for content surfaces.
 export const SheetContainer = React.lazy(() => import('./SheetContainer'));
