+++ packages/plugins/plugin-sheet/src/model/index.ts
@@ -2,5 +2,6 @@
 // Copyright 2024 DXOS.org
 //
 
+export * from './decorations';
 export * from './formatting-model';
 export * from './sheet-model';
